DWR-1496 Moved things around in preparation for a different embargo SQL for export/pdf vs view exams
