+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/StudentHistoryExamWrapper.java
@@ -1,7 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.student;
-
-import org.opentestsystem.rdw.reporting.common.model.Assessment;
-import org.opentestsystem.rdw.reporting.common.model.Exam;
+package org.opentestsystem.rdw.reporting.common.model;
 
 /**
  * This model represents an Exam within the context of a student's history.
@@ -27,6 +24,10 @@ public class StudentHistoryExamWrapper {
      */
     public Assessment getAssessment() {
         return assessment;
+    }
+
+    public static Builder builder() {
+        return new Builder();
     }
 
     public static class Builder {
