+++ common/src/test/resources/application-test-current-school-year.yml
