+++ report-processor/src/main/resources/application.sql.yml
@@ -17,7 +17,7 @@ sql:
         AND (:grade_id is null or a.grade_id=:grade_id)
         AND a.type_id = 2
         AND ${sql.reportProcessor.snippet.schoolFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
       ORDER BY e.completed_at DESC,
         i.performance_task_writing_type DESC
 
@@ -41,7 +41,7 @@ sql:
         AND (:grade_id is null or a.grade_id=:grade_id)
         AND a.type_id=2
         AND ${sql.reportProcessor.snippet.schoolFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
       ORDER BY e.completed_at DESC,
               i.performance_task_writing_type DESC
 
@@ -57,7 +57,7 @@ sql:
         AND (:grade_id is null or a.grade_id=:grade_id)
         AND a.type_id=1
         AND ${sql.reportProcessor.snippet.schoolFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
       ORDER BY e.completed_at asc
 
     findAllIcasByStudentIdsAndQueryParamsGroupPermissions: >-
@@ -76,7 +76,7 @@ sql:
         AND (:grade_id is null or a.grade_id=:grade_id)
         AND a.type_id=1
         AND ${sql.reportProcessor.snippet.schoolFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
       ORDER BY e.completed_at asc
 
   export:
@@ -198,7 +198,7 @@ sql:
         JOIN school isch ON st.inferred_school_id = isch.id
       WHERE
         ${sql.reportProcessor.snippet.exportExamsFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
 
     hasExams: >-
       SELECT 1
@@ -212,7 +212,7 @@ sql:
         JOIN school isch ON st.inferred_school_id = isch.id
       where
         ${sql.reportProcessor.snippet.exportExamsFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
 
   student:
     findByExamQueryParamsUnknownPermissions: >-
@@ -223,7 +223,7 @@ sql:
         AND (:grade_id is null or a.grade_id = :grade_id)
         AND (:assessment_type is null or a.type_id = :assessment_type)
         AND ${sql.reportProcessor.snippet.schoolFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
       group by st.id
 
     findByExamQueryParamsGroupPermissions: >-
@@ -238,11 +238,22 @@ sql:
         AND (:grade_id is null or a.grade_id = :grade_id)
         AND (:assessment_type is null or a.type_id = :assessment_type)
         AND ${sql.reportProcessor.snippet.schoolFilter}
-        AND ${sql.snippet.studentExamPermission}
+        AND ${sql.reportProcessor.snippet.studentExamPermission}
       GROUP BY st.id
 
   reportProcessor:
     snippet:
+      studentExamPermission: >-
+            ${sql.snippet.studentExamPermissionNoEmbargo}
+            AND
+            (
+               e.school_year < ${app.school-year}
+               OR e.type_id <> 3
+               OR s.embargo_enabled = 0
+               OR 1 = :state_embargo_admin
+               OR s.district_id IN (:district_embargo_admin_ids)
+            )
+
       schoolFilter:
             (
               :school_id is null
