+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/repository/JdbcExamEmbargoPermissionCurrentSchoolYearIT.java
@@ -1,37 +1,53 @@
-package org.opentestsystem.rdw.reporting.exam.student;
+package org.opentestsystem.rdw.reporting.processor.repository;
 
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.common.model.AssessmentType;
 import org.opentestsystem.rdw.common.model.Subject;
-import org.opentestsystem.rdw.reporting.JdbcRepositoryWithPermissionIT;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
 import org.opentestsystem.rdw.reporting.common.model.ScaleScore;
 import org.opentestsystem.rdw.reporting.common.model.School;
 import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
+import org.opentestsystem.rdw.reporting.common.test.ITDataSourceConfiguration;
+import org.opentestsystem.rdw.reporting.common.test.RepositoryIT;
 import org.opentestsystem.rdw.reporting.common.test.support.Instants;
+import org.opentestsystem.rdw.reporting.common.web.security.User;
+import org.opentestsystem.rdw.security.Permission;
+import org.opentestsystem.rdw.security.PermissionScope;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlGroup;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.opentestsystem.rdw.common.model.AssessmentType.SUMMATIVE;
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.getEmbargoPermissions;
+import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.individualOf;
+import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.permissions;
 
 
 /**
  * Test the embargo related SQL clause of the permissions.
  * This is done by configuring the school-year to the test data value
  */
+@RunWith(SpringRunner.class)
+@RepositoryIT
 @SuppressWarnings({"SqlNoDataSourceInspection", "SqlResolve"})
-@Import(JdbcStudentExamRepository.class)
 @ActiveProfiles({"test", "test-current-school-year"})
+@ContextConfiguration(classes = {ITDataSourceConfiguration.class, JdbcStudentExamPermissionRepository.class})
 @SqlGroup({
         @Sql(scripts = {"classpath:integration-test-data-permissions.sql"}),
         @Sql(statements = {"update school set embargo_enabled = 0 where id <> -30",
@@ -47,10 +63,11 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
                         " (-99, 1, -1, 'g1', -1, -30, 0, 0, 0, 0, 0, 1997, -1, 'v1', 2, 'Complete', 1, 'Valid', 'session1', 2000, 20, 1, '1997-01-01 00:00:00.000000', 1, 100, 10, 2, 200, 20, 3, 300, 30, 4, 400, 40, -1, '1997-07-18 20:14:34.000000', -1)"
         })
 })
-public class JdbcExamEmbargoPermissionCurrentSchoolYearIT extends JdbcRepositoryWithPermissionIT {
+public class JdbcExamEmbargoPermissionCurrentSchoolYearIT {
+
 
     @Autowired
-    private StudentExamRepository repository;
+    private StudentExamPermissionRepository repository;
 
     @Autowired
     JdbcTemplate template;
@@ -58,8 +75,17 @@ public class JdbcExamEmbargoPermissionCurrentSchoolYearIT extends JdbcRepository
     private int examCount;
     private int sumExamSchool30Count;
 
+    private final static Map<String, Permission> individualStatewide = ImmutableMap.copyOf(permissions(individualOf(PermissionScope.STATEWIDE)));
+    private User.Builder userBuilder;
+
+
     @Before
     public void setUp() {
+        userBuilder = User.builder()
+                .id(UUID.randomUUID().toString())
+                .username("someone-10@somewhere.com")
+                .password("redacted");
+
         sumExamSchool30Count = template.queryForObject("SELECT count(*) FROM exam e " +
                 "JOIN asmt a ON a.id = e.asmt_id " +
                 "JOIN school s ON s.id = e.school_id" +
