+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/repository/JdbcStudentExamPermissionRepository.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.exam.student;
+package org.opentestsystem.rdw.reporting.processor.repository;
 
 import com.google.common.collect.ImmutableMap;
 import org.opentestsystem.rdw.reporting.common.jdbc.Assessments;
@@ -6,12 +6,10 @@ import org.opentestsystem.rdw.reporting.common.jdbc.Exams;
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
-import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
 import org.opentestsystem.rdw.reporting.common.security.PermissionSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
@@ -19,20 +17,17 @@ import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @Repository
-class JdbcStudentExamRepository implements StudentExamRepository {
+class JdbcStudentExamPermissionRepository implements StudentExamPermissionRepository {
 
     private final NamedParameterJdbcTemplate template;
     private final SecurityParameterProvider securityParameterProvider;
 
-    @Value("${sql.exam.findAllByStudentId}")
+    @Value("${sql.examPermissionTest.findAllByStudentId}")
     private String findAllByStudentIdQuery;
 
-    @Value("${sql.exam.existsForStudentId}")
-    private String existsForStudentIdQuery;
-
     @Autowired
-    JdbcStudentExamRepository(final NamedParameterJdbcTemplate template,
-                              final SecurityParameterProvider securityParameterProvider) {
+    JdbcStudentExamPermissionRepository(final NamedParameterJdbcTemplate template,
+                                        final SecurityParameterProvider securityParameterProvider) {
         this.template = template;
         this.securityParameterProvider = securityParameterProvider;
     }
@@ -51,27 +46,6 @@ class JdbcStudentExamRepository implements StudentExamRepository {
                             .exam(Exams.map(row, Exam.builder(), permissionSource.getPermissionsById()).build())
                             .assessment(Assessments.map(row, Assessment.builder(), "asmt_").build())
                             .build();
-                }
-        );
-    }
-
-    @Override
-    public Student findStudentWithExams(@NotNull final PermissionSource permissionSource, final String ssid) {
-        try {
-            return template.queryForObject(
-                    existsForStudentIdQuery,
-                    ImmutableMap.<String, Object>builder()
-                            .putAll(securityParameterProvider.getSecurityParameters(permissionSource))
-                            .put("student_ssid", ssid)
-                            .build(),
-                    (row, index) -> Student.builder()
-                            .id(row.getLong("id"))
-                            .ssid(ssid)
-                            .firstName(row.getString("first_name"))
-                            .lastName(row.getString("last_or_surname"))
-                            .build());
-        } catch (final EmptyResultDataAccessException exception) {
-            return null;
-        }
+                });
     }
 }
