+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/repository/StudentExamPermissionRepository.java
@@ -1,13 +1,15 @@
-package org.opentestsystem.rdw.reporting.exam.student;
+package org.opentestsystem.rdw.reporting.processor.repository;
 
-import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
 import org.opentestsystem.rdw.reporting.common.security.PermissionSource;
 
 import javax.validation.constraints.NotNull;
 import java.util.List;
 
-interface StudentExamRepository {
+/**
+ * Used in permission related test
+ */
+interface StudentExamPermissionRepository {
 
     /**
      * Find all of the exams allowed to view for the given {@link PermissionSource}.
@@ -16,13 +18,4 @@ interface StudentExamRepository {
      * @param studentId        The student SSID  @return  The student's exam history
      */
     List<StudentHistoryExamWrapper> findAllForStudent(@NotNull PermissionSource permissionSource, long studentId);
-
-
-    /**
-     * Return the student if at least one exam exists.
-     *
-     * @param permissionSource the security information to prepare the SQL query parameters for
-     * @param ssid             The student SSID  @return The student
-     */
-    Student findStudentWithExams(@NotNull PermissionSource permissionSource, String ssid);
 }
