+++ reporting-service/src/main/java/org/opentestsystem/rdw/reporting/exam/student/DefaultStudentExamService.java
@@ -1,6 +1,7 @@
 package org.opentestsystem.rdw.reporting.exam.student;
 
 import org.opentestsystem.rdw.reporting.common.model.Student;
+import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -50,5 +51,4 @@ class DefaultStudentExamService implements StudentExamService {
 
         return student;
     }
-
 }
