+++ reporting-service/src/main/java/org/opentestsystem/rdw/reporting/exam/student/StudentExamHistory.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.reporting.exam.student;
 
 import com.google.common.collect.ImmutableSet;
 import org.opentestsystem.rdw.reporting.common.model.Student;
+import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
 
 import java.util.Collection;
