+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/JdbcRepositoryWithPermissionIT.java
@@ -26,7 +26,6 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 @RunWith(SpringRunner.class)
 @RepositoryIT
 @ContextConfiguration(classes = {ITDataSourceConfiguration.class})
-@Sql(scripts = {"classpath:integration-test-data.sql"})
 @ActiveProfiles("test")
 public abstract class JdbcRepositoryWithPermissionIT {
