+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/assessmentgrade/JdbcAssessmentGradeRepositoryIT.java
@@ -5,6 +5,7 @@ import org.opentestsystem.rdw.reporting.JdbcRepositoryWithPermissionIT;
 import org.opentestsystem.rdw.reporting.exam.Grade;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.individualOf;
@@ -12,6 +13,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.schools;
 
 @Import(JdbcAssessmentGradeRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcAssessmentGradeRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
