+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/group/JdbcGroupAssessmentRepositoryIT.java
@@ -8,6 +8,7 @@ import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.security.GroupGrant;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.groupOf;
@@ -15,6 +16,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.schools;
 
 @Import(JdbcGroupAssessmentRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcGroupAssessmentRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
