+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/group/JdbcGroupExamItemRepositoryIT.java
@@ -7,6 +7,7 @@ import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.security.GroupGrant;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import java.util.Collection;
 import java.util.List;
@@ -17,6 +18,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.schools;
 
 @Import(JdbcGroupExamItemRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcGroupExamItemRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
