+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/schoolgrade/JdbcSchoolGradeAssessmentRepositoryIT.java
@@ -7,6 +7,7 @@ import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.individualOf;
@@ -14,6 +15,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.schools;
 
 @Import(JdbcSchoolGradeAssessmentRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcSchoolGradeAssessmentRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
