+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/schoolgrade/JdbcSchoolGradeExamItemRepositoryIT.java
@@ -6,6 +6,7 @@ import org.opentestsystem.rdw.reporting.common.model.ExamItem;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import java.util.Collection;
 
@@ -13,6 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 
 @Import(JdbcSchoolGradeExamItemRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcSchoolGradeExamItemRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
