+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/schoolgrade/JdbcSchoolGradeExamRepositoryIT.java
@@ -7,6 +7,7 @@ import org.opentestsystem.rdw.reporting.common.model.Exam;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.groupOf;
@@ -16,6 +17,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 
 
 @Import(JdbcSchoolGradeExamRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcSchoolGradeExamRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
