+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/student/DefaultStudentExamServiceTest.java
@@ -8,6 +8,7 @@ import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
 import org.opentestsystem.rdw.reporting.common.model.Student;
+import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.security.Permission;
 import org.opentestsystem.rdw.security.PermissionScope;
