+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/student/JdbcExamItemRepositoryIT.java
@@ -5,6 +5,7 @@ import org.opentestsystem.rdw.reporting.JdbcRepositoryWithPermissionIT;
 import org.opentestsystem.rdw.reporting.common.model.ExamItem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import java.util.List;
 
@@ -16,6 +17,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 
 
 @Import(JdbcExamItemRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcExamItemRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
