+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/student/JdbcExamPermissionIT.java
@@ -8,6 +8,7 @@ import org.opentestsystem.rdw.security.PermissionScope;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.context.jdbc.Sql;
 
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -29,13 +30,14 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
  */
 @SuppressWarnings({"SqlNoDataSourceInspection", "SqlResolve"})
 @Import(JdbcStudentExamRepository.class)
+@Sql(scripts = {"classpath:integration-test-data-permissions.sql"})
 public class JdbcExamPermissionIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
     private StudentExamRepository repository;
 
     @Autowired
-    JdbcTemplate template;
+    private JdbcTemplate template;
 
     private int examCount;
     private int mathExamCount;
