+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/student/JdbcStudentExamRepositoryIT.java
@@ -6,8 +6,10 @@ import org.opentestsystem.rdw.reporting.TestData;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
 import org.opentestsystem.rdw.reporting.common.model.Student;
+import org.opentestsystem.rdw.reporting.common.model.StudentHistoryExamWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.test.context.jdbc.Sql;
 
 import java.util.List;
 
@@ -19,6 +21,7 @@ import static org.opentestsystem.rdw.reporting.common.test.support.PermissionSco
 import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.schools;
 
 @Import(JdbcStudentExamRepository.class)
+@Sql(scripts = {"classpath:integration-test-data.sql"})
 public class JdbcStudentExamRepositoryIT extends JdbcRepositoryWithPermissionIT {
 
     @Autowired
